uipath_etl/requirements.txt
requests
mysql-connector-python
pandas
bcrypt
flask
flask-cors
pyjwt
flasgger
